flow repaired

+++ Assets/Scripts/Dto/RequestDto.cs
@@ -6,7 +6,7 @@ public class RequestDto
 {
 	public int version;
 	public float size;
-	public float taskId;
+	public int taskId;
 	public string graphName;
 	public ActionsDto actions;
 	public List<FilterDto> filters;
